feat() : get move filter model to external file - move Filter model to external model. - create FitechBook service to call any api of books. feat #feature/filter_books
