+++ src/services/books/FetchBook.js
@@ -1,0 +1,11 @@
+export const filterBooks = (
+  price,
+  array,
+  arabic,
+  english,
+  french,
+  spanish,
+  reviewChecked,
+) => {
+  return 'Price : ' + price + '\nCategories : ' + array;
+};
